Improved UI functionality
- Clone object from context menu
- Remove object from context menu
- Crash on removing entity with opened list fixed
- Uncomplete graphics component copying fixed
- Uncomplete animation component copying fixed
- Event system refactored
- Crash on removing selected object fixed

+++ Game/src/Core/EntityComponents/Physics.cpp
@@ -46,7 +46,7 @@ namespace Oblivion
 
 			this->mBody->CreateFixture(&shape, density);
 
-			this->mBody->SetTransform(b2Vec2((float)pos.x, (float)pos.y), 0);
+			this->mBody->SetTransform(b2Vec2(pos.x, pos.y), 0);
 
 			break;
 		}
@@ -58,7 +58,6 @@ namespace Oblivion
 	{
 		mBody->GetWorld()->DestroyBody(mBody);
 	}
-
 
 	void PhysicsComponent::Update(const float& ts)
 	{
